Convert to browser-native ES6 import syntax and limit Electron API usage
