+++ frontend/components/sidebar.js
@@ -431,6 +431,10 @@ export class Sidebar extends LitElement {
           a.classList.add('is-selected')
         })
       })
+
+      const gettingStarted = (this.shadowRoot ?? this).querySelector("#getting_starting_tab");
+      gettingStarted.click()
+
   }
 
   render() {
