+++ frontend/electron/check.js
@@ -1,0 +1,3 @@
+var userAgent = navigator.userAgent.toLowerCase();
+
+export const isElectron = userAgent.indexOf(' electron/') > -1
