+++ frontend/electron/shell-commands.js
@@ -1,5 +1,4 @@
 const shell = require("electron").shell;
-
 const links = document.querySelectorAll("a[href]");
 
 Array.prototype.forEach.call(links, (link) => {
@@ -24,22 +23,11 @@ function directToDocumentation() {
     document.getElementById("overview-column-2").blur();
     // window.open('https://docs.sodaforsparc.io', '_blank');
 }
-const directToGuidedMode = () => {
-    const guidedModeLinkButton = document.getElementById("guided_mode_view");
-    guidedModeLinkButton.click();
-};
-const directToFreeFormMode = () => {
-    const freeFormModeLinkButton = document.getElementById("main_tabs_view");
-    freeFormModeLinkButton.click();
-};
 
 document.getElementById("doc-btn").addEventListener("click", directToDocumentation);
 document
     .getElementById("home-button-interface-instructions-link")
     .addEventListener("click", sodaVideo);
-document
-    .getElementById("home-button-guided-mode-link")
-    .addEventListener("click", directToGuidedMode);
-document
-    .getElementById("home-button-free-form-mode-link")
-    .addEventListener("click", directToFreeFormMode);
+
+
+exports.default = shell;
