+++ frontend/index.js
@@ -2,3 +2,6 @@
 // Web Components
 import "./components/sidebar.js";
 import "./components/multiselect/index.js";
+
+import "./nav.js" // Activate navigation behaviors
+import "./renderer.js"
